remove sentPacketHandler from packetPacker
